Split out VFS name resolver into separate file.
