+++ sys/kern/Makefile
@@ -61,7 +61,7 @@ SOURCES = \
 	uio.c \
 	ustack.c \
 	vfs.c \
-	vfs_lookup.c \
+	vfs_name.c \
 	vfs_readdir.c \
 	vfs_syscalls.c \
 	vfs_vnode.c \
